New Leaders and such
United States is now a faction. The Federal Territories are under the control of the Federal Gov and the states as typical pre 1936. The autonomous territories are slightly more, well, autonomous, but still answer to the Federal Government. Will need to add neat events and whatnot to give the feeling of participating in the new american politics.